Love-8/Love-task8-task-202411081445/gen_src/Deck.java
File: Deck.java
```java
import java.util.List;

public class Deck {
    private List<Card> cards;
    private static Deck instance;
    
    private Deck() {
        // constructor logic
    }
    
    public static Deck getInstance() {
        // method logic
    }
    
    public void shuffle() {
        // method logic
    }
    
    public Card dealCard() {
        // method logic
    }
}
```

File: Card.java
```java
public class Card {
    private String rank;
    private String suit;
    
    public Card(String rank, String suit) {
        // constructor logic
    }
    
    public String getRank() {
        // method logic
    }
    
    public String getSuit() {
        // method logic
    }
}
```